// src/store/governmentHolidays/governmentHolidays.enum.ts
enum HolidayTypeEnum {
  GET_HOLIDAY_PENDING = "GET_HOLIDAY_PENDING",
  GET_HOLIDAY_SUCCESS = "GET_HOLIDAY_SUCCESS",
  GET_HOLIDAY_FAILED = "GET_HOLIDAY_FAILED",

  GET_ALL_VENDORS_PENDING = "GET_ALL_VENDORS_PENDING",
  GET_ALL_VENDORS_SUCCESS = "GET_ALL_VENDORS_SUCCESS",
  GET_ALL_VENDORS_FAILED = "GET_ALL_VENDORS_FAILED",

  ADD_HOLIDAY_PENDING = "ADD_HOLIDAY_PENDING",
  ADD_HOLIDAY_SUCCESS = "ADD_HOLIDAY_SUCCESS",
  ADD_HOLIDAY_FAILED = "ADD_HOLIDAY_FAILED",

  UPDATE_HOLIDAY_PENDING = "UPDATE_HOLIDAY_PENDING",
  UPDATE_HOLIDAY_SUCCESS = "UPDATE_HOLIDAY_SUCCESS",
  UPDATE_HOLIDAY_FAILED = "UPDATE_HOLIDAY_FAILED",

  DELETE_HOLIDAY_PENDING = "DELETE_HOLIDAY_PENDING",
  DELETE_HOLIDAY_SUCCESS = "DELETE_HOLIDAY_SUCCESS",
  DELETE_HOLIDAY_FAILED = "DELETE_HOLIDAY_FAILED",
}
export default HolidayTypeEnum;